requirements.txt
matplotlib==3.5.2
numpy==1.21.0
opencv_python==4.5.2.54
pandas==1.4.3
Pillow==9.2.0
pyTelegramBotAPI==4.6.1
python-dotenv==0.20.0
requests==2.25.1
skimage==0.0
tensorflow==2.9.1
torch==1.12.0
torchvision==0.13.0
